# apidocumentation.yaml
openapi: 3.0.2
info:
  description: This API is used with the Million Song Dataset made available by the CORGIS
    dataset project to provide users with information related to songs and their
    affiliated artists. The purpose of the API is to deliver to users both basic
    and advanced features based on this dataset through a Web app.
  version: 1.0.14
  title: Million Songs API
  termsOfService: https://corgis-edu.github.io/corgis/csv/music/
  contact:
    email: group25@rug.nl
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  /song/search:
    get:
      tags:
        - Song
      description: Filter through songs using artist, terms, or year.
      parameters:
      - name: artist
        in: query
        description: Filter songs by artist
        schema:
          type: string
          example: Drake
      - name: term
        in: query
        description: Filter songs by terms
        schema:
          type: string
          example: Hip Hop
      - name: year
        in: query
        description: Filter songs by year
        schema:
          type: integer
          example: 2017
      - name: top
        in: query
        description: Size of the top
        schema:
          type: integer
          minimum: 1
          maximum: 10000
          example: 2
      responses:
        200:
          description: Successfull pull of artists information from the database
          content:
            application/json:
              schema:
                 $ref: '#/components/schemas/Song'
              example:
                - artist_mbtags: 0.0
                  artist_mbtags_count: 0.0
                  bars_confidence: 0.806
                  bars_start: 1.91886
                  beats_confidence: 0.44
                  beats_start: 1.22595
                  duration: 269.81832
                  end_of_fade_in: 5.3
                  hotttnesss: 0.265861049
                  id: "SOIAZJW12AB01853F1"
                  key: 4
                  key_confidence: 0.717
                  loudness: -13.496
                  mode: 1
                  mode_confidence: 0.652
                  start_of_fade_out: 258.99
                  tatums_confidence: 0.121
                  tatums_start: 1.22595
                  tempo: 86.643
                  time_signature: 4.0
                  time_signature_confidence: 0.487
                  title: 0.0
                  year: 1984
                - artist_mbtags: 0.0
                  artist_mbtags_count: 0.0
                  bars_confidence: 0.018
                  bars_start: 1.45439
                  beats_confidence: 1.0
                  beats_start: 0.42896
                  duration: 245.21098
                  end_of_fade_in: 0.357
                  hotttnesss: 0.265861049
                  id: "SOKEJEJ12A8C13E0D0"
                  key: 7
                  key_confidence: 0.07
                  loudness: -7.545
                  mode: 1
                  mode_confidence: 0.686
                  start_of_fade_out: 227.48
                  tatums_confidence: 0.728
                  tatums_start: 0.1727
                  tempo: 117.975
                  time_signature: 4.0
                  time_signature_confidence: 0.835
                  title: 0.0
                  year: 0.0
            text/csv:
              schema:
                 $ref: '#/components/schemas/Song'
              example:
                artist_mbtags,artist_mbtags_count,bars_confidence,bars_start,beats_confidence,beats_start,duration,end_of_fade_in,hotttnesss,id,key,key_confidence,loudness,mode,mode_confidence,start_of_fade_out,tatums_confidence,tatums_start,tempo,time_signature,time_signature_confidence,title,year

                0,0,0.806,1.91886,0.44,1.22595,269.81832,5.3,0.265861049,SOIAZJW12AB01853F1,4,0.717,-13.496,1,0.652,258.99,0.121,1.22595,86.643,4,0.487,0,1984

                0,0,0.018,1.45439,1,0.42896,245.21098,0.357,0.265861049,SOKEJEJ12A8C13E0D0,7,0.07,-7.545,1,0.686,227.48,0.728,0.1727,117.975,4,0.835,0,0
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The request was unacceptable, often due to missing a required parameter.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The requested resource doesn't exist.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The requested resource doesn't exist."

        500:
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Server error.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: "Server error."
...
  /artist/search:
    get:
      tags:
        - Artist
      description:  Obtain a list of all artists in the dataset, optionally filtered by artist name and/or genre.
      parameters:
      - name: name
        in: query
        required: false
        description: Used to filter artistst by artist name
        schema:
          type: string
          example: "MxPx"
      - name: terms
        in: query
        required: false
        description: Used to filter artists by genre
        schema:
          type: string
          example: "christian rock"
      responses:
        200:
          description: Succesfully pulled the list of artists from the database.
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Song'
                example:
                  - artist_mbtags: 0.0
                    artist_mbtags_count: 2.0
                    bars_confidence: 1.0
                    bars_start: 0.10403
                    beats_confidence: 1.0
                    beats_start: 0.10403
                    duration: 151.24853
                    end_of_fade_in: 0.312
                    hotttnesss: 0.466304831
                    id: "SOLCNJN12A6D4F7740"
                    key: 11.0
                    key_confidence: 0.0
                    loudness: -5.41
                    mode: 0
                    mode_confidence: 0.108
                    start_of_fade_out: 141.891
                    tatums_confidence: 0.551
                    tatums_start: 0.10403
                    tempo: 101.106
                    time_signature: 1.0
                    time_signature_confidence: 0.0
                    title: 0.0
                    year: 1995
                  -
                    artist_mbtags: 0.0
                    artist_mbtags_count: 2.0
                    bars_confidence: 0.021
                    bars_start: 0.37448
                    beats_confidence: 0.227
                    beats_start: 0.37448
                    duration: 215.24853
                    end_of_fade_in: 0.0
                    hotttnesss: -1
                    id: "SOMOWDB12AF72A8AE2"
                    key: 5.0
                    key_confidence: 0.854
                    loudness: -5.316
                    mode: 1
                    mode_confidence: 0.496
                    start_of_fade_out: 205.264
                    tatums_confidence: 0.753
                    tatums_start: 0.37448
                    tempo: 171.693
                    time_signature: 4.0
                    time_signature_confidence: 0.928
                    title: 0.0
                    year: 1995
              text/csv:
                schema:
                  $ref: '#/components/schemas/Song'
                example:
                  artist_mbtags,artist_mbtags_count,bars_confidence,bars_start,beats_confidence,beats_start,duration,end_of_fade_in,hotttnesss,id,key,key_confidence,loudness,mode,mode_confidence,start_of_fade_out,tatums_confidence,tatums_start,tempo,time_signature,time_signature_confidence,title,year

                  0,2,1,0.10403,1,0.10403,151.24853,0.312,0.466304831,SOLCNJN12A6D4F7740,11,0,-5.41,0,0.108,141.891,0.551,0.10403,101.106,1,0,0,1995

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The request was unacceptable, often due to missing a required parameter.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."
        402:
          description: Request Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The parameters were valid but the request failed.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The parameters were valid but the request failed."

        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The requested resource doesn't exist.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The requested resource doesn't exist."

        500:
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Server error.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "Server error."
...
  /rating:
    post:
      tags:
        - Rating
      description: Create the database for rankings.
      responses:
        200:
          description: Successfully created the rating databse
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Successfully created the database for ratings
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The request was unacceptable, often due to missing a required parameter.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The requested resource doesn't exist.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The requested resource doesn't exist."

        500:
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Server error.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "Server error."


## - Rating an song - ##
    put:
      tags:
        - Rating
      description: Send the rating for a song or an artist.
      parameters:
        - in: query
          name: song
          description: The unique ID of a song
          schema:
            type: string
            example: '"SOMZWCG12A8C13C480"'
        - in: query
          name: artist
          description: The unique ID of an artist
          schema:
            type: string
            example: '"ARR3ONV1187B9A2F59"'
      responses:
        200:
          description: "Sending a new rating of a song or artists"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Successfull rating sending
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                  "The request was unacceptable, often due to missing a required parameter."

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The request was unacceptable, often due to missing a required parameter.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The requested resource doesn't exist.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The requested resource doesn't exist."

        500:
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Server error.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "Server error."



## - Rating an artist - ##
    delete:
      tags:
        - Rating
      description: Delete the rating for an artist or song.
      parameters:
        - in: query
          name: song
          description: The unique ID of a song
          schema:
            type: string
            example: '"SOMZWCG12A8C13C480"'
        - in: query
          name: artist
          description: The unique ID of an artist
          schema:
            type: string
            example: '"ARR3ONV1187B9A2F59"'
      responses:
        200:
          description: "Delete the rating for an artist or song"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Successfull rating sending
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                  "The request was unacceptable, often due to missing a required parameter."

        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The request was unacceptable, often due to missing a required parameter.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The request was unacceptable, often due to missing a required parameter."
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: The requested resource doesn't exist.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "The requested resource doesn't exist."

        500:
          description: Something went wrong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response: Server error.
            text/csv:
              schema:
                $ref: '#/components/schemas/API_Response'
              example:
                response

                "Server error."





# externalDocs:
#   description: Find out more about Swagger
#   url: http://swagger.io
servers:
  - url: https://localhost:3001/
  - url: http://localhost:3001/
components:
  schemas:
...
    Song:
      type: object
      properties:
        artist_mbtags:
          type: number
          format: float
        artist_mbtags_count:
          type: number
          format: float
          description: Number of tags for the artist on mbtags
        bars_confidence:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: Confidence value associated with each bar
        bars_start:
          type: number
          format: float
          description: Average start time of each bar, measured in bars
        beats_confidence:
          type: number
          format: float
          description: Average confidence interval of the beats
        beats_start:
          type: number
          format: float
          description: Average start time of each beat, measured in beats
        duration:
          type: number
          format: float
          description: Duration of the track in seconds.
        end_of_fade_in:
          type: number
          format: float
          description: Time of the end of the fade in, at the beginning of the song.
        hotttnesss:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: A measure of the song's popularity
        id:
          type: string
          description: A uniquely identifying number for the song.
        key:
          type: number
          minimum: 0
          maximum: 11
          format: float
          description: Estimation of the key the song is in
        key_confidence:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: Confidence value of the key estimation
        loudness:
          type: number
          format: float
          description: General loudness of the track
        mode:
          type: integer
          format: int32
          description: Estimation of the mode the song
        mode_confidence:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: Confidence value of the mode estimation
        start_of_fade_out:
          type: number
          format: float
          description: Start time of the fade out, in seconds, at the end of the song
        tatums_confidence:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: Confidence value associated with each tatum
        tatums_start:
          type: number
          format: float
          description: Average start time of each tatum, measured in tatums
        tempo:
          type: number
          format: float
          description: Tempo in BPM
        time_signature:
          type: number
          format: float
          description: Time signature of the song, i.e. usual number of beats per bar.
        time_signature_confidence:
          type: number
          minimum: 0
          maximum: 1
          format: float
          description: Confidence of the time signature estimation
        title:
          type: integer
          format: int64
          description: Name of the song
        year:
          type: integer
          format: int32
          description: Year when this song was released
      xml:
        name: Song
...
    API_Response:
      type: object
      properties:
        response:
          type: string
          description: Response from the server